+++ resources/js/bootstrap/cp.ts
@@ -21,6 +21,7 @@ import {setUrlDefaults} from '@/wayfinder';
 import {inertiaPageRegistry, resolveInertiaPage} from './inertia-pages.js';
 import AppLayout from '@/common/layouts/AppLayout.vue';
 import {createCpComponentRegistry} from './components.js';
+import {registerSlideoutGlobals} from '@/common/slideouts';
 import {registerFormComponents} from '@/modules/forms/register';
 import {configureIcons} from './icons.js';
 
@@ -161,6 +162,7 @@ const Cp = {
 
     handleNonInertiaRequests();
     ensureLegacyNotificationContainer();
+    registerSlideoutGlobals();
 
     console.log('Calling booted callbacks', bootedCallbacks);
     bootedCallbacks.forEach((callback) => callback(this));